Extract pitr agent's logic into separate file
